Multiple clients support
Now is possible to handle multiple socket connections.
Also the core was refactored and coverage test was added.
