+++ package.json
@@ -1,12 +1,11 @@
 {
   "name": "wsreq",
-  "version": "1.1.2",
-  "description": "Websocket testing made easy, even with weird events.",
+  "version": "2.0.1",
+  "description": "Socket.io endpoints testing made easy.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development ts-node-dev --respawn example",
-    "dev:client": "cross-env NODE_ENV=development ts-node-dev --respawn example/client.ts",
+    "dev": "cross-env NODE_ENV=development ts-node-dev --respawn src/example",
     "test": "cross-env NODE_ENV=test jest --runInBand",
     "build": "tsc"
   },
@@ -15,14 +14,16 @@
     "url": "git+https://github.com/Aku-mi/wsreq.git"
   },
   "keywords": [
-    "websockets",
-    "request",
-    "rest",
+    "nodejs",
+    "http",
+    "socket-io",
+    "websocket",
+    "requests",
     "api",
-    "express",
-    "socket.io",
     "test",
-    "testing"
+    "testing",
+    "http-server",
+    "rest-api"
   ],
   "author": "Aku-mi",
   "license": "MIT",
