--- pipeline.yml
@@ -1,16 +1,0 @@
-trigger:
-  - master
-
-pool:
-  vmImage: ubuntu-latest
-
-steps:
-  - task: NodeTool@0
-    inputs:
-      versionSpec: "14.x"
-    displayName: "Install Node.js"
-
-  - script: |
-      npm install
-      npm run test
-    displayName: "npm install and test"
